requirements.txt
# FastAPI and server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Document Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
python-docx>=0.8.11
pandas>=2.0.0

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0

# Embeddings
sentence-transformers>=2.2.0

# Vector Store
faiss-cpu>=1.7.4

# LLM Providers (choose one or more)
groq>=0.4.0
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.18.0

# Utilities
numpy>=1.24.0
pydantic>=2.0.0

# Optional: Web UI
streamlit>=1.28.0

# Development
pytest>=7.4.0
httpx>=0.25.0

# Windows-specific (for async)
python-dotenv>=1.0.0
